finalized version

+++ backend/routes/api/reviews.js
@@ -145,6 +145,15 @@ router.put('/:reviewId', authenticateUser, async (req, res, next) => {
 
 router.delete('/:reviewId', authenticateUser, async (req, res, next) => {
     const { reviewId } = req.params;
+
+    const currentReview = await Review.findByPk(reviewId);
+    if(!currentReview) {
+        const err = new Error("Review couldn't be found");
+        err.title = "Review couldn't be found";
+        err.errors = ["Review couldn't be found"];
+        err.status = 404;
+        return next(err);
+    };
     await Review.destroy({where: { id:reviewId }});
 
     res.json ({
